Added mlb.result package with Match,TeamHistory,YearRange
Match, YearRange , TeamHistory  case classes  and the creation of a JSON encoder (JsonEncoder) for it within the mlb.result package. 
To create a JSON encoder for the case classes, the DeriveJsonEncoder feature of the zio-json library is used. The DeriveJsonEncoder.gen method automatically generates a JSON encoder for a class.
when the objects are converted to JSON format, this encoder is used.